backend/project/requirements.txt
arrow>=1.2.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
blessed==1.19.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
crispy-bootstrap5==0.6
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-dynamic-fixture==3.1.2
django-environ==0.4.5
django-extensions==3.1.3
django-jsoneditor==0.1.6
django-leaflet==0.28.1
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.2
nodeenv==1.6.0
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.0
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==2.4.7
pyperclip==1.8.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.7
wcwidth==0.2.5